The HTTP client timeout does not meet all the expectations it should. It acts both as a connect and idle timeout.
Introduce separate connect and idle timeouts for HTTP client, the connect timeout fires when a request has not been obtained from the pool in time, the idle timeout is a rename of the existing timeout (on HttpClientRequest).

RequestOptions deprecates the timeout property and introduces a connect timeout and an idle timeout. To achieve the same effect both connect/idle timeouts can be set.

HttpClientRequest deprecates the setTimeout method in favor of setIdleTimeout which is a better and explicit name.
